refactor(queries): colocate advance settings queries
